Move PCAMX to a separate file
